:tada: 프로젝트 진행을 위한 메인 소스코드 초기화
