+++ packages/three/vite.config.ts
@@ -9,11 +9,4 @@ export default defineConfig({
       '@': path.resolve(__dirname, './src'),
     },
   },
-  css: {
-    preprocessorOptions: {
-      scss: {
-        path: path.resolve(__dirname, './scss'),
-      },
-    },
-  },
 });
